Marketing_updates/requirements.txt
azure-ai-projects~=1.0.0b7
azure-identity>=1.12.0
cryptography>=41.0.0
asyncio>=3.4.3
python-dotenv>=1.0.0
aiohttp>=3.8.5
argparse>=1.4.0
typing-extensions>=4.7.0 